Collect all custom reusable widgets in the widgets library.
