+++ app/lib/conversation_pane/conversation_content/text_message_tile.dart
@@ -5,9 +5,9 @@
 import 'package:flutter/material.dart';
 import 'package:prototype/conversation_pane/message_renderer.dart';
 import 'package:prototype/core/core.dart';
-import 'package:prototype/elements.dart';
 import 'package:prototype/styles.dart';
 import 'package:prototype/user_cubit.dart';
+import 'package:prototype/widgets/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:prototype/core_extension.dart';
