+++ app/lib/conversation_pane/conversation_details/add_members.dart
@@ -4,10 +4,10 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:prototype/elements.dart';
 import 'package:prototype/navigation/navigation.dart';
 import 'package:prototype/styles.dart';
 import 'package:prototype/user_cubit.dart';
+import 'package:prototype/widgets/widgets.dart';
 
 import 'add_members_cubit.dart';
 
@@ -43,7 +43,7 @@ class AddMembersView extends StatelessWidget {
         backgroundColor: Colors.white,
         elevation: 0,
         scrolledUnderElevation: 0,
-        leading: appBarBackButton(context),
+        leading: const AppBarBackButton(),
         title: const Text("Add members"),
       ),
       body: Padding(
