+++ app/lib/registration/username_password.dart
@@ -6,9 +6,9 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:prototype/elements.dart';
 import 'package:prototype/navigation/navigation.dart';
 import 'package:prototype/styles.dart';
+import 'package:prototype/widgets/widgets.dart';
 
 import 'registration_cubit.dart';
 
@@ -22,7 +22,7 @@ class UsernamePasswordChoice extends StatelessWidget {
       appBar: AppBar(
         title: const Text('Sign up', style: TextStyle(fontFamily: fontFamily)),
         toolbarHeight: isPointer() ? 100 : null,
-        leading: appBarBackButton(context),
+        leading: const AppBarBackButton(),
       ),
       body: Padding(
         padding: const EdgeInsets.all(20.0),
