+++ app/lib/widgets/app_bar_back_button.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+// SPDX-FileCopyrightText: 2025 Phoenix R&D GmbH <hello@phnx.im>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+class AppBarBackButton extends StatelessWidget {
+  const AppBarBackButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      icon: const Icon(Icons.arrow_back),
+      color: Colors.black,
+      hoverColor: Colors.transparent,
+      splashColor: Colors.transparent,
+      highlightColor: Colors.transparent,
+      onPressed: () => Navigator.of(context).pop(),
+    );
+  }
+}
