+++ app/lib/widgets/frosted_glass.dart
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: 2025 Phoenix R&D GmbH <hello@phnx.im>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+import 'dart:ui';
+
+import 'package:flutter/material.dart';
+
+class FrostedGlass extends StatelessWidget {
+  const FrostedGlass({
+    super.key,
+    required this.color,
+    required this.height,
+  });
+
+  final Color color;
+  final double height;
+
+  @override
+  Widget build(BuildContext context) {
+    return ClipRect(
+      child: BackdropFilter(
+        filter: ImageFilter.blur(
+            sigmaX: 15, sigmaY: 15, tileMode: TileMode.repeated),
+        child: Container(
+          width: MediaQuery.of(context).size.width,
+          height: height,
+          color: color.withValues(alpha: 0.4),
+        ),
+      ),
+    );
+  }
+}
