+++ app/lib/widgets/widgets.dart
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2025 Phoenix R&D GmbH <hello@phnx.im>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+/// Collection of custom reusable widgets
+library;
+
+export 'app_bar_back_button.dart';
+export 'user_avatar.dart';
+export 'frosted_glass.dart';
